Move ProblemsDetailFragment to fragments package
